reafctored slack stuff into new notification file
